% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_clip.R
\name{signal_clip}
\alias{signal_clip}
\title{Clip signal based on time vector.}
\usage{
signal_clip(data, time, limits)
}
\arguments{
\item{data}{\code{eseis} object, \code{numeric} vector or list of 
objects, data set to be processed.}

\item{time}{\code{POSIXct} vector, corresponding time vector. Only needed 
if \code{data} is no \code{eseis} object.}

\item{limits}{\code{POSIXct} vector of length two, time limits for 
clipping.}
}
\value{
\code{Numeric} data set clipped to provided time interval.
}
\description{
The function clips a seismic signal based on the corresponding time vector.
}
\examples{

print("Not included in Lite version")

}
\author{
Michael Dietze
}
\keyword{eseis}
