% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_deconvolve.R
\name{signal_deconvolve}
\alias{signal_deconvolve}
\title{Deconvolve a signal vector.}
\usage{
signal_deconvolve(
  data,
  sensor = "TC120s",
  logger = "Cube3BOB",
  gain = 1,
  use_metadata = FALSE,
  dt,
  p = 10^-6,
  waterlevel = 10^-6,
  na.replace = FALSE
)
}
\arguments{
\item{data}{\code{eseis} object, \code{numeric} vector or list of 
objects, data set to be processed.}

\item{sensor}{\code{Character} value or \code{list} object, 
seismic sensor name. Must be present in the sensor library 
(\code{list_sensor}) or parameters must be added manually (see 
examples). Default is \code{"TC120s"}.}

\item{logger}{\code{Character} value, seismic logger name. Must be 
present in the logger library (\code{list_logger}) or parameters must be
added manually. Default is \code{"Cube3extBOB"}.}

\item{gain}{\code{Numeric} value, signal gain level of the logger. Default 
is \code{1}.}

\item{use_metadata}{\code{Logical} value, option to take keywords for 
\code{sensor}, \code{logger} and \code{gain} from eseis object meta data
element instead of using explicitly provided arguments. Default is 
\code{FALSE}.}

\item{dt}{\code{Numeric} value, sampling rate. Only needed if \code{data} 
is not an \code{eseis} object}

\item{p}{\code{Numeric} value, proportion of signal to be tapered. Default
is\code{10^-6}.}

\item{waterlevel}{\code{Numeric} value, waterlevel value for frequency
division, default is \code{10^-6}.}

\item{na.replace}{\code{Logical} value, option to replace NA values in the 
data set by zeros. Default is \code{FALSE}. Attention, the zeros will 
create artifacts in the deconvolved data set. However, \code{NA} values will 
result in no deconvolution at all.}
}
\value{
\code{Numeric} vector or list of vectors, deconvolved signal.
}
\description{
The function removes the instrument response from a signal vector.
}
\details{
The function requires a set of input parameters, apart from the signal 
vector. These parameters are contained in and read from the function 
\code{list_sensor()} and \code{list_logger()}. Poles and zeros are used 
to build the transfer function. The value s is the generator constant in 
Vs/m. The value k is the normalisation factor. AD is the analogue-digital 
conversion factor. If the signal was recorded with a gain value other than 
1, the resulting signal needs to be corrected for this, as well.
}
\examples{

print("Not included in Lite version")

}
\author{
Michael Dietze
}
\keyword{eseis}
