% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_demean.R
\name{signal_demean}
\alias{signal_demean}
\title{Remove mean of signal vector.}
\usage{
signal_demean(data)
}
\arguments{
\item{data}{\code{eseis} object, \code{numeric} vector or list of 
objects, data set to be processed.}
}
\value{
\code{Numeric} vector or list of vectors, data set with mean 
subtracted.
}
\description{
The function removes the mean from a signal vector.
}
\examples{

print("Not included in Lite version")

}
\author{
Michael Dietze
}
\keyword{eseis}
