% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_sum.R
\name{signal_sum}
\alias{signal_sum}
\title{Calculate signal vector sum.}
\usage{
signal_sum(...)
}
\arguments{
\item{\dots}{\code{Numeric} vectors or \code{eseis} objects, input 
signal, that must be of the same length.}
}
\value{
\code{Numeric} vector, signal vector sum.
}
\description{
The function calculates the vector sum of the input signals.
}
\examples{

print("Not included in Lite version")

}
\author{
Michael Dietze
}
\keyword{eseis}
