% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_taper.R
\name{signal_taper}
\alias{signal_taper}
\title{Taper a signal vector.}
\usage{
signal_taper(data, p = 0, n)
}
\arguments{
\item{data}{\code{eseis} object, \code{numeric} vector or list of 
objects, data set to be processed.}

\item{p}{\code{Numeric} value, proportion of the signal vector to be 
tapered. Alternative to \code{n}.}

\item{n}{\code{Numeric} value, number of samples to be tapered at each
end of the signal vector.}
}
\value{
\code{Data frame}, tapered signal vector.
}
\description{
The function tapers a signal vector with a cosine bell taper, either of a
given proportion or a discrete number of samples.
}
\examples{

print("Not included in Lite version")

}
\author{
Michael Dietze
}
\keyword{eseis}
