% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_aggregate.R
\name{time_aggregate}
\alias{time_aggregate}
\title{Aggregate a time series}
\usage{
time_aggregate(data, n = 2)
}
\arguments{
\item{data}{\code{POSIXct} vector, time to be processed.}

\item{n}{\code{Numeric} value, number of samples to be aggregated to one
new data value. Must be an integer value greater than 1. Default is 
\code{2}.}
}
\value{
\code{POSIXct} vector, aggregated data.
}
\description{
The time series \code{x} is aggregated by an integer factor \code{n}.
}
\examples{

print("Not included in Lite version")

}
\author{
Michael Dietze
}
\keyword{eseis}
