% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_clip.R
\name{time_clip}
\alias{time_clip}
\title{Clip time vector.}
\usage{
time_clip(time, limits)
}
\arguments{
\item{time}{\code{POSIXct} vector, time vector.}

\item{limits}{\code{POSIXct} vector of length two, time limits for 
clipping.}
}
\value{
\code{POSIXct} vector, clipped time vector.
}
\description{
The function clips a time vector based on provided limits.
}
\examples{

print("Not included in Lite version")

}
\author{
Michael Dietze
}
\keyword{eseis}
